backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg[binary]
alembic
pydantic
twilio
requests
psycopg2-binary
python-multipart   # required for Twilio form-data webhook
loguru              # for better logging
